del old files

+++ detect_meteor/util.py
@@ -4,9 +4,10 @@ from logzero import logger
 
 
 def safe_os_remove(full_path):
-    try:
-        os.remove(full_path)
-    except:
-        msg = traceback.format_exc()
-        logger.warn("error when try remove: " + msg)
+    if os.path.exists(full_path):
+        try:
+            os.remove(full_path)
+        except:
+            msg = traceback.format_exc()
+            logger.warn("error when try remove: " + msg)
